requirements.txt
Django
django-chunked-upload
djangorestframework
mysqlclient
python-magic-bin